requirements.txt
flask
python-binance
pandas
python-dotenv
textblob
snscrape
psutil
pytz
requests
numpy
scipy
scikit-learn>=1.3.0
joblib>=1.3.0
TA-Lib
ccxt
schedule>=1.2.0
supabase>=2.0.0
openpyxl>=3.1.0